requirements.txt
# VLM OCR Document Reader - Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pymupdf>=1.23.0
PyYAML>=6.0.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Logging
python-json-logger>=2.0.0
